refactor(search): Move sealing and deletion-blocker logic into cl/search/deletion_utils.py
Sealing business logic (RECAPDocument's seal_documents/delete_from_ia,
the OpinionCluster file cleanup and delete-plus-redirect flow, and the
deletion-blocker checks that gate it) was split between
cl/search/utils.py and OpinionClusterAdmin itself. The admin isn't a
good place for it to live: pull it all into one dedicated module so
it's easy to find and doesn't require an admin instance to invoke.

- cl/search/deletion_utils.py: new home for delete_from_ia,
  seal_documents, delete_cluster_files, seal_cluster (holding the
  cluster-delete/ClusterRedirection/docket-delete logic previously
  inlined in OpinionClusterAdmin.seal_cluster), and the
  SEAL_BLOCKERS_MAP/CLUSTER_BLOCKER_KEYS/DOCKET_BLOCKER_KEYS constants
  plus check_blocking_relations/get_blocking_relations/
  get_deletion_blockers, previously methods on OpinionClusterAdmin.
- cl/search/admin.py: seal_cluster_view, blocking_confirmation_view,
  and seal_clusters now call the deletion_utils functions directly;
  the admin methods are gone.
- cl/lib/admin.py, the seal_documents and delete_document_from_ia
  management commands, and tests updated to import from the new
  module.
